# sbtc-bridge-api/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"KeySet": {
				"properties": {
					"deposits": {
						"properties": {
							"oraclePubKey": {
								"type": "string"
							},
							"reclaimPubKey": {
								"type": "string"
							},
							"revealPubKey": {
								"type": "string"
							}
						},
						"required": [
							"oraclePubKey",
							"reclaimPubKey",
							"revealPubKey"
						],
						"type": "object"
					}
				},
				"required": [
					"deposits"
				],
				"type": "object"
			},
			"Uint8Array": {
				"description": "A typed array of 8-bit unsigned integer values. The contents are initialized to 0. If the\r\nrequested number of bytes could not be allocated an exception is raised.",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"PeginScriptI": {
				"properties": {
					"tweakedPubkey": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"tapMerkleRoot": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"tapLeafScript": {},
					"tapInternalKey": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"leaves": {},
					"wsh": {
						"type": "string"
					},
					"witnessScript": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"redeemScript": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"paymentType": {
						"type": "string"
					},
					"script": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Uint8Array"
							}
						]
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"paymentType"
				],
				"type": "object"
			},
			"depositPayloadType": {
				"properties": {
					"amountSats": {
						"type": "number",
						"format": "double"
					},
					"revealFee": {
						"type": "number",
						"format": "double"
					},
					"memo": {
						"type": "string"
					},
					"lengthOfMemo": {
						"type": "number",
						"format": "double"
					},
					"cname": {
						"type": "string"
					},
					"lengthOfCname": {
						"type": "number",
						"format": "double"
					},
					"stacksAddress": {
						"type": "string"
					},
					"prinType": {
						"type": "number",
						"format": "double"
					},
					"opcode": {
						"type": "string"
					}
				},
				"required": [
					"amountSats",
					"revealFee",
					"lengthOfMemo",
					"lengthOfCname",
					"stacksAddress",
					"prinType",
					"opcode"
				],
				"type": "object"
			},
			"withdrawalPayloadType": {
				"properties": {
					"dustAmount": {
						"type": "number",
						"format": "double"
					},
					"amountSats": {
						"type": "number",
						"format": "double"
					},
					"compression": {
						"type": "number",
						"format": "double"
					},
					"signature": {
						"type": "string"
					},
					"stacksAddress": {
						"type": "string"
					},
					"opcode": {
						"type": "string"
					}
				},
				"required": [
					"amountSats",
					"compression",
					"signature",
					"stacksAddress",
					"opcode"
				],
				"type": "object"
			},
			"FeeEstimateResponse": {
				"properties": {
					"feeInfo": {
						"properties": {
							"high_fee_per_kb": {
								"type": "number",
								"format": "double"
							},
							"medium_fee_per_kb": {
								"type": "number",
								"format": "double"
							},
							"low_fee_per_kb": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"high_fee_per_kb",
							"medium_fee_per_kb",
							"low_fee_per_kb"
						],
						"type": "object"
					}
				},
				"required": [
					"feeInfo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BalanceI": {
				"properties": {
					"balance": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"balance"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddressValidationI": {
				"properties": {
					"witness_program": {
						"type": "string"
					},
					"witness_version": {
						"type": "number",
						"format": "double"
					},
					"iswitness": {
						"type": "boolean"
					},
					"isscript": {
						"type": "boolean"
					},
					"scriptPubKey": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"isvalid": {
						"type": "boolean"
					}
				},
				"required": [
					"witness_program",
					"witness_version",
					"iswitness",
					"isscript",
					"scriptPubKey",
					"address",
					"isvalid"
				],
				"type": "object"
			},
			"SbtcContractDataI": {
				"properties": {
					"addressValidation": {
						"$ref": "#/components/schemas/AddressValidationI"
					},
					"burnHeight": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					},
					"totalSupply": {
						"type": "number",
						"format": "double"
					},
					"threshold": {
						"type": "number",
						"format": "double"
					},
					"tokenUri": {
						"type": "string"
					},
					"tradingHalted": {
						"type": "boolean"
					},
					"numParties": {
						"type": "number",
						"format": "double"
					},
					"numKeys": {
						"type": "number",
						"format": "double"
					},
					"sbtcWalletPublicKey": {
						"type": "string"
					},
					"sbtcWalletAddress": {
						"type": "string"
					},
					"contractOwner": {
						"type": "string"
					},
					"coordinator": {
						"properties": {
							"key": {
								"type": "string"
							},
							"addr": {
								"properties": {
									"value": {
										"type": "string"
									}
								},
								"required": [
									"value"
								],
								"type": "object"
							}
						},
						"required": [
							"key",
							"addr"
						],
						"type": "object"
					}
				},
				"required": [
					"sbtcWalletPublicKey",
					"sbtcWalletAddress",
					"contractOwner"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "bridge-api",
		"version": "1.0.0",
		"description": "Enables caching of contract data for the sBTC Bridge app",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/bridge-api/{network}/v1/btc/tx/keys": {
			"get": {
				"operationId": "GetKeys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KeySet"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/btc/tx/commitment/{stxAddress}/{revealFee}": {
			"get": {
				"operationId": "Commitment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PeginScriptI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "stxAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "revealFee",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/tx/commit-deposit-data/{stxAddress}/{revealFee}": {
			"get": {
				"operationId": "CommitDepositData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "stxAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "revealFee",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/tx/commit-deposit/{data}": {
			"get": {
				"operationId": "CommitDeposit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/depositPayloadType"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "data",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/tx/commit-withdrawal-data/{signature}/{amount}": {
			"get": {
				"operationId": "CommitWithdrawalData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "signature",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "amount",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/tx/commit-withdrawal/{data}/{sbtcWallet}/{compression}": {
			"get": {
				"operationId": "CommitWithdrawal",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/withdrawalPayloadType"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "data",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "sbtcWallet",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "compression",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/tx/{txid}": {
			"get": {
				"operationId": "FetchRawTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/tx/{txid}/hex": {
			"get": {
				"operationId": "FetchTransactionHex",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/wallet/address/{address}/txs": {
			"get": {
				"operationId": "FetchAddressTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/wallet/loadwallet/{name}": {
			"get": {
				"operationId": "LoadWallet",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/btc/wallet/listwallets": {
			"get": {
				"operationId": "ListWallets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/btc/blocks/fee-estimate": {
			"get": {
				"operationId": "GetFeeEstimate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FeeEstimateResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/btc/blocks/info": {
			"get": {
				"operationId": "GetInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/btc/blocks/count": {
			"get": {
				"operationId": "GetCount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/sbtc/pegins/scan": {
			"get": {
				"operationId": "ScanPeginRequests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/sbtc/pegins/commits/scan/{btcAddress}/{stxAddress}/{sbtcWalletAddress}/{revealFee}": {
			"get": {
				"operationId": "ScanCommitments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "btcAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "stxAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "sbtcWalletAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "revealFee",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/sbtc/address/{address}/balance": {
			"get": {
				"operationId": "FetchUserSbtcBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BalanceI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bridge-api/{network}/v1/sbtc/data": {
			"get": {
				"operationId": "FetchSbtcContractData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SbtcContractDataI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/sbtc/wallet-address": {
			"get": {
				"operationId": "FetchSbtcWalletAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/config": {
			"get": {
				"operationId": "GetAllParam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/bridge-api/{network}/v1/config/{param}": {
			"get": {
				"operationId": "GetParam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"type": "number",
											"format": "double"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "param",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}